requirements.txt
langchain-ollama
langgraph
langgraph-checkpoint-sqlite
langchain-community
duckduckgo-search
gradio
fpdf2